Moves views to MacPlatform to better reuse
