+++ main/src/addins/MacPlatform/AssemblyInfo.cs
@@ -8,3 +8,4 @@ using System.Runtime.CompilerServices;
 [assembly: AssemblyDescription ("Mac Platform Support for MonoDevelop")]
 [assembly: AssemblyVersion ("2.6")]
 [assembly: AssemblyCopyright ("MIT X11")]
+[assembly: InternalsVisibleTo ("MonoDevelop.DesignerSupport, PublicKey=002400000c800000940000000602000000240000525341310004000001000100e1290d741888d13312c0cd1f72bb843236573c80158a286f11bb98de5ee8acc3142c9c97b472684e521ae45125d7414558f2e70ac56504f3e8fe80830da2cdb1cda8504e8d196150d05a214609234694ec0ebf4b37fc7537e09d877c3e65000f7467fa3adb6e62c82b10ada1af4a83651556c7d949959817fed97480839dd39b")]
